requirements.txt
asttokens==3.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
fonttools==4.61.0
idna==3.11
ImageIO==2.37.2
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kagglehub==0.3.13
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.7
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.3
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
stack-data==0.6.3
tifffile==2025.5.10
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
